# ast/term: don't sort an object's keys slice in-place