--- a/test/cases/testdata/partialsetdoc/test-issue-3819.yaml
+++ b/test/cases/testdata/partialsetdoc/test-issue-3819.yaml
@@ -1,0 +1,24 @@
+cases:
+- data:
+    a:
+      bananas: 1
+      apples: 2
+      clementines: 3
+  modules:
+  - |
+    package foo
+    p[x] {
+        data.a[x] # iterates data.a
+        q[_]
+    }
+    q[x] {
+        x = data.a # inserts data.a into a set, sorts the object keys
+    }
+  note: 'partialsetdoc: object sort while iter'
+  query: data.foo.p = x
+  sort_bindings: true
+  want_result:
+  - x:
+    - apples
+    - bananas
+    - clementines