Two different commits, two different approaches to fix the underlying problem that during evaluation, we may sort an object's key slice; and that could happen within the function passed to the object's `Iter()` method, causing the weird situation that on subsequent invocations of the passed function, the *same key* would be presented.

This example shows how it's possible, and I believe that this is what happened in #3819.

The partial set memo work that seemed to be the cause of the problem must have changed one thing: before, the document under `data.external["test.target"]` would _not_ have been Compare'd to anything, and afterwards, it would. (I haven't uncovered the complete code flow yet...)

----

This was introduced in https://github.com/open-policy-agent/opa/commit/2381824b168f0d776ec3fc9454e9b002b97fc036, and there's a TODO comment hinting at this being not ideal. I suppose we want to contemplate which way to go here (squash and merge, or drop the second commit doing the sort at insertion-time).

----

Fixes #3819.