@srenatus 2021-09-27
Running some benchmarks with this:

As is to be expected, `(obj).String()` (which used to sort before printing) got faster, since it's got a lot less to do:
```
name                               old time/op    new time/op    delta
ObjectString/5/String()-16            893ns ± 7%     564ns ± 2%  -36.78%  (p=0.008 n=5+5)
ObjectString/50/String()-16          10.4µs ± 2%     4.6µs ± 5%  -55.76%  (p=0.008 n=5+5)
ObjectString/500/String()-16          170µs ± 3%      58µs ± 2%  -66.09%  (p=0.008 n=5+5)
ObjectString/5000/String()-16        2.40ms ± 6%    0.60ms ± 5%  -74.90%  (p=0.008 n=5+5)

name                               old alloc/op   new alloc/op   delta
ObjectString/5/String()-16             280B ± 0%      208B ± 0%  -25.71%  (p=0.008 n=5+5)
ObjectString/50/String()-16          2.26kB ± 0%    1.82kB ± 0%  -19.43%  (p=0.008 n=5+5)
ObjectString/500/String()-16         41.8kB ± 0%    37.6kB ± 0%   -9.87%  (p=0.008 n=5+5)
ObjectString/5000/String()-16         409kB ± 0%     368kB ± 0%  -10.01%  (p=0.008 n=5+5)

name                               old allocs/op  new allocs/op  delta
ObjectString/5/String()-16             21.0 ± 0%      19.0 ± 0%   -9.52%  (p=0.008 n=5+5)
ObjectString/50/String()-16             159 ± 0%       157 ± 0%   -1.26%  (p=0.008 n=5+5)
ObjectString/500/String()-16          1.52k ± 0%     1.51k ± 0%   -0.13%  (p=0.008 n=5+5)
ObjectString/5000/String()-16         11.0k ± 0%     11.0k ± 0%   -0.02%  (p=0.008 n=5+5)
```

This is "paid for" by extra work during object creation, like the added benchmark shows:

```
name                         old time/op    new time/op    delta
ObjectConstruction/5-16        4.11µs ± 1%    4.79µs ± 7%   +16.59%  (p=0.008 n=5+5)
ObjectConstruction/50-16       36.6µs ± 2%    46.8µs ±10%   +27.99%  (p=0.008 n=5+5)
ObjectConstruction/500-16       412µs ± 7%     557µs ± 3%   +35.09%  (p=0.008 n=5+5)
ObjectConstruction/5000-16     4.80ms ± 6%    7.65ms ± 2%   +59.47%  (p=0.008 n=5+5)
ObjectConstruction/50000-16    52.7ms ± 6%   182.6ms ± 7%  +246.27%  (p=0.008 n=5+5)

name                         old alloc/op   new alloc/op   delta
ObjectConstruction/5-16        2.97kB ± 0%    2.97kB ± 0%      ~     (all equal)
ObjectConstruction/50-16       19.8kB ± 0%    19.8kB ± 0%      ~     (p=0.357 n=5+5)
ObjectConstruction/500-16       244kB ± 0%     244kB ± 0%      ~     (p=0.222 n=5+5)
ObjectConstruction/5000-16     2.24MB ± 0%    2.24MB ± 0%      ~     (p=0.095 n=5+5)
ObjectConstruction/50000-16    23.4MB ± 1%    23.8MB ± 0%    +1.71%  (p=0.008 n=5+5)

name                         old allocs/op  new allocs/op  delta
ObjectConstruction/5-16          65.0 ± 0%      65.0 ± 0%      ~     (all equal)
ObjectConstruction/50-16          556 ± 0%       556 ± 0%      ~     (all equal)
ObjectConstruction/500-16       5.53k ± 0%     5.53k ± 0%      ~     (p=1.000 n=5+5)
ObjectConstruction/5000-16      55.1k ± 0%     55.1k ± 0%      ~     (p=1.000 n=4+5)
ObjectConstruction/50000-16      553k ± 0%      553k ± 0%      ~     (p=1.000 n=5+5)
```